pyproject.toml
[project]
name = "blog_ve"
version = "0.1.0"
description = "blog_dv"
authors = [
    {name = "nemo", email = "60929499+youyou1cha@users.noreply.github.com"},
]
dependencies = [
    "django>=5.0.6",
    "django-cors-headers>=4.3.1",
    "djangorestframework>=3.15.1",
    "django-filter>=24.2",
    "yapf>=0.40.2",
    "markdown>=3.6",
    "Pillow>=10.3.0",
    "rest-framework-simplejwt>=0.0.2",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
